requirements.txt
accelerate==1.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
bitsandbytes-cuda110==0.26.0.post2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
h11==0.14.0
haystack-ai==2.6.1
haystack-experimental==0.2.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
kiwisolver==1.4.7
lazy-imports==0.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openai==1.51.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
posthog==3.7.0
psutil==6.0.0
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
qi==3.1.5
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
sentencepiece==0.2.0
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
starlette==0.40.0
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
validators==0.34.0
watchfiles==0.24.0
websockets==13.1
